+++ include/sprite_loader.h
