+++ src/stb_image.cpp
@@ -1,2 +1,2 @@
 #define STB_IMAGE_IMPLEMENTATION
-#include "headers/stb_image.h"
+#include "stb_image.h"
